Remove session controller override and add to dedicated extend
