--- app/controllers/decidim/devise/sessions_controller.rb
@@ -1,67 +1,0 @@
-# frozen_string_literal: true
-
-module Decidim
-  module Devise
-    # Custom Devise SessionsController to avoid namespace problems.
-    class SessionsController < ::Devise::SessionsController
-      include Decidim::DeviseControllers
-
-      # rubocop: disable Rails/LexicallyScopedActionFilter
-      before_action :check_sign_in_enabled, only: :create
-      # rubocop: enable Rails/LexicallyScopedActionFilter
-
-      def destroy
-        current_user.invalidate_all_sessions!
-        if session["omniauth.france_connect.end_session_uri"].present?
-          destroy_france_connect_session(session["omniauth.france_connect.end_session_uri"])
-        elsif params[:translation_suffix].present?
-          super { set_flash_message! :notice, params[:translation_suffix], { scope: "decidim.devise.sessions" } }
-        else
-          super
-        end
-      end
-
-      def after_sign_in_path_for(user)
-        if user.present? && user.blocked?
-          check_user_block_status(user)
-        elsif first_login_and_not_authorized?(user) && !user.admin? && !pending_redirect?(user)
-          decidim_verifications.first_login_authorizations_path
-        else
-          super
-        end
-      end
-
-      # Calling the `stored_location_for` method removes the key, so in order
-      # to check if there's any pending redirect after login I need to call
-      # this method and use the value to set a pending redirect. This is the
-      # only way to do this without checking the session directly.
-      def pending_redirect?(user)
-        store_location_for(user, stored_location_for(user))
-      end
-
-      def first_login_and_not_authorized?(user)
-        user.is_a?(User) && user.sign_in_count == 1 && current_organization.available_authorizations.any? && user.verifiable?
-      end
-
-      def after_sign_out_path_for(user)
-        request.referer || super
-      end
-
-      private
-
-      def check_sign_in_enabled
-        redirect_to new_user_session_path unless current_organization.sign_in_enabled?
-      end
-
-      def destroy_france_connect_session(fc_logout_path)
-        signed_out = (::Devise.sign_out_all_scopes ? sign_out : sign_out(resource_name))
-        if signed_out
-          set_flash_message! :notice, :signed_out
-          session.delete("omniauth.france_connect.end_session_uri")
-        end
-
-        redirect_to fc_logout_path
-      end
-    end
-  end
-end
